Refactor test example and create prerelease nuget
