+++ src/Xchain.Tests/Test02.cs
@@ -1,22 +1,30 @@
 
 namespace Xchain.Tests;
 
+[CollectionDefinition("Second")]
+public class SecondCollection : 
+    // This collection will start only after FirstTest collection is done
+    ICollectionFixture<Test01AwaiterFixture>, 
+    ICollectionFixture<CollectionChainFixture>;
 
-[Collection("Third")]
-public class _03_LastTest(CollectionChainFixture chain)
+
+[Collection("Second")]
+public class Test02(CollectionChainFixture chain)
 {
 
     [Fact()]
     public void LinkedTest1() => chain.Link((output) =>
     {
+        chain.Output["x"] = 10;
         Thread.Sleep(5000);
         throw new NotImplementedException();
     });
 
 
     [Fact()]
-    public void LinkedTest22() => chain.Link((output) =>
+    public void LinkedTest2() => chain.Link((output) =>
     {
+        chain.Output["x"] = 10;
         Thread.Sleep(5000);
         throw new NotImplementedException();
     });
@@ -24,6 +32,7 @@ public class _03_LastTest(CollectionChainFixture chain)
     [Fact()]
     public void LinkedTest3() => chain.Link((output) =>
     {
+        chain.Output["x"] = 10;
         Thread.Sleep(5000);
     });
 
@@ -31,13 +40,15 @@ public class _03_LastTest(CollectionChainFixture chain)
     [Fact()]
     public void LinkedTest4() => chain.Link((output) =>
     {
+        var sleep = output.Get("Sleep");
         Thread.Sleep(5000);
-        throw new NotImplementedException(output.Get("Sleep"));
+        throw new NotImplementedException(sleep);
     });
 
     [Fact()]
     public void LinkedTest5() => chain.Link((output) =>
     {
+        chain.Output["x"] = 10;
         Thread.Sleep(5000);
     });
 }
